unit tests added, CheckoutServices function moved to separate file
