+++ .eslintignore
@@ -11,7 +11,7 @@ src/cartridges/int_adyen_SFRA/cartridge/client/default/js/bundle.js
 # Extended SFRA core file
 src/cartridges/int_adyen_SFRA/cartridge/client/default/js/checkout/checkoutSFRA5.js
 src/cartridges/int_adyen_SFRA/cartridge/client/default/js/checkout/checkoutSFRA6.js
-src/cartridges/int_adyen_SFRA/cartridge/controllers/CheckoutServices.js
+src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/checkout_services/placeOrder.js
 
 # Test files
 src/cartridges/int_adyen_SFRA/test
