+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/CheckoutServices.js
@@ -1,234 +1,9 @@
-'use strict';
-
 const server = require('server');
+
 server.extend(module.superModule);
 
-/* ### Custom Adyen cartridge ### */
-const adyenHelpers = require('*/cartridge/scripts/checkout/adyenHelpers');
-const constants = require('*/cartridge/adyenConstants/constants');
-const { processPayment, isNotAdyen } = require('*/cartridge/controllers/middlewares/checkout_services/adyenCheckoutServices');
-/* ### Custom Adyen cartridge ### */
+const placeOrder = require('*/cartridge/controllers/middlewares/checkout_services/placeOrder');
 
-server.prepend('PlaceOrder', server.middleware.https, function (req, res, next) {
-  const BasketMgr = require('dw/order/BasketMgr');
-  const OrderMgr = require('dw/order/OrderMgr');
-  const Resource = require('dw/web/Resource');
-  const Transaction = require('dw/system/Transaction');
-  const URLUtils = require('dw/web/URLUtils');
-  const basketCalculationHelpers = require('*/cartridge/scripts/helpers/basketCalculationHelpers');
-  const hooksHelper = require('*/cartridge/scripts/helpers/hooks');
-  const COHelpers = require('*/cartridge/scripts/checkout/checkoutHelpers');
-  const validationHelpers = require('*/cartridge/scripts/helpers/basketValidationHelpers');
-  const addressHelpers = require('*/cartridge/scripts/helpers/addressHelpers');
-
-  const currentBasket = BasketMgr.getCurrentBasket();
-
-  if (!currentBasket) {
-    res.json({
-      error: true,
-      cartError: true,
-      fieldErrors: [],
-      serverErrors: [],
-      redirectUrl: URLUtils.url('Cart-Show').toString()
-    });
-    return next();
-  }
-
-  /* ### Custom Adyen cartridge ### */
-  if(isNotAdyen(currentBasket)) {
-    return next();
-  }
-  /* ### Custom Adyen cartridge ### */
-  var validatedProducts = validationHelpers.validateProducts(currentBasket);
-  if (validatedProducts.error) {
-    res.json({
-      error: true,
-      cartError: true,
-      fieldErrors: [],
-      serverErrors: [],
-      redirectUrl: URLUtils.url('Cart-Show').toString()
-    });
-    return next();
-  }
-
-  if (req.session.privacyCache.get('fraudDetectionStatus')) {
-    res.json({
-      error: true,
-      cartError: true,
-      redirectUrl: URLUtils.url('Error-ErrorCode', 'err', '01').toString(),
-      errorMessage: Resource.msg('error.technical', 'checkout', null)
-    });
-
-    return next();
-  }
-
-  var validationOrderStatus = hooksHelper('app.validate.order', 'validateOrder', currentBasket, require('*/cartridge/scripts/hooks/validateOrder').validateOrder);
-  if (validationOrderStatus.error) {
-    res.json({
-      error: true,
-      errorMessage: validationOrderStatus.message
-    });
-    return next();
-  }
-
-  // Check to make sure there is a shipping address
-  if (currentBasket.defaultShipment.shippingAddress === null) {
-    res.json({
-      error: true,
-      errorStage: {
-        stage: 'shipping',
-        step: 'address'
-      },
-      errorMessage: Resource.msg('error.no.shipping.address', 'checkout', null)
-    });
-    return next();
-  }
-
-  // Check to make sure billing address exists
-  if (!currentBasket.billingAddress) {
-    res.json({
-      error: true,
-      errorStage: {
-        stage: 'payment',
-        step: 'billingAddress'
-      },
-      errorMessage: Resource.msg('error.no.billing.address', 'checkout', null)
-    });
-    return next();
-  }
-
-  // Calculate the basket
-  Transaction.wrap(function () {
-    basketCalculationHelpers.calculateTotals(currentBasket);
-  });
-
-  // Re-validates existing payment instruments
-  var validPayment = COHelpers.validatePayment(req, currentBasket);
-  if (validPayment.error) {
-    res.json({
-      error: true,
-      errorStage: {
-        stage: 'payment',
-        step: 'paymentInstrument'
-      },
-      errorMessage: Resource.msg('error.payment.not.valid', 'checkout', null)
-    });
-    return next();
-  }
-
-  // Re-calculate the payments.
-  var calculatedPaymentTransactionTotal = COHelpers.calculatePaymentTransaction(currentBasket);
-  if (calculatedPaymentTransactionTotal.error) {
-    res.json({
-      error: true,
-      errorMessage: Resource.msg('error.technical', 'checkout', null)
-    });
-    return next();
-  }
-
-  // Creates a new order.
-  var order = COHelpers.createOrder(currentBasket);
-  if (!order) {
-    res.json({
-      error: true,
-      errorMessage: Resource.msg('error.technical', 'checkout', null)
-    });
-    return next();
-  }
-
-  // Cache order number in order to be able to restore cart later
-  req.session.privacyCache.set('currentOrderNumber', order.orderNo);
-
-  /* ### Custom Adyen cartridge ### */
-  // Handles payment authorization
-  var handlePaymentResult = adyenHelpers.handlePayments(order, order.orderNo);
-  /* ### Custom Adyen cartridge ### */
-
-  // Handle custom processing post authorization
-  var options = {
-    req: req,
-    res: res
-  };
-  var postAuthCustomizations = hooksHelper('app.post.auth', 'postAuthorization', handlePaymentResult, order, options, require('*/cartridge/scripts/hooks/postAuthorizationHandling').postAuthorization);
-  if (postAuthCustomizations && Object.prototype.hasOwnProperty.call(postAuthCustomizations, 'error')) {
-    res.json(postAuthCustomizations);
-    return next();
-  }
-
-  if (handlePaymentResult.error) {
-    res.json({
-      error: true,
-      errorMessage: Resource.msg('error.payment.not.valid', 'checkout', null)
-    });
-    this.emit('route:Complete', req, res);
-    return;
-  }
-
-  /* ### Custom Adyen cartridge ### */
-  const cbEmitter = (route) => this.emit(route, req, res);
-  if (
-      handlePaymentResult.action &&
-      handlePaymentResult.action?.type !== constants.ACTIONTYPES.VOUCHER
-  ) {
-    return processPayment(order, handlePaymentResult, req, res, cbEmitter);
-  }
-  /* ### Custom Adyen cartridge ### */
-
-  var fraudDetectionStatus = hooksHelper('app.fraud.detection', 'fraudDetection', currentBasket, require('*/cartridge/scripts/hooks/fraudDetection').fraudDetection);
-  if (fraudDetectionStatus.status === 'fail') {
-    Transaction.wrap(function () { OrderMgr.failOrder(order, true); });
-
-    // fraud detection failed
-    req.session.privacyCache.set('fraudDetectionStatus', true);
-
-    res.json({
-      error: true,
-      cartError: true,
-      redirectUrl: URLUtils.url('Error-ErrorCode', 'err', fraudDetectionStatus.errorCode).toString(),
-      errorMessage: Resource.msg('error.technical', 'checkout', null)
-    });
-
-    return next();
-  }
-
-  // Places the order
-  var placeOrderResult = COHelpers.placeOrder(order, fraudDetectionStatus);
-
-  if (placeOrderResult.error) {
-    res.json({
-      error: true,
-      errorMessage: Resource.msg('error.technical', 'checkout', null)
-    });
-    return next();
-  }
-
-  if (req.currentCustomer.addressBook) {
-    // save all used shipping addresses to address book of the logged in customer
-    var allAddresses = addressHelpers.gatherShippingAddresses(order);
-    allAddresses.forEach(function (address) {
-      if (!addressHelpers.checkIfAddressStored(address, req.currentCustomer.addressBook.addresses)) {
-        addressHelpers.saveAddress(address, req.currentCustomer, addressHelpers.generateAddressName(address));
-      }
-    });
-  }
-
-  if (order.getCustomerEmail()) {
-    COHelpers.sendConfirmationEmail(order, req.locale.id);
-  }
-
-  // Reset usingMultiShip after successful Order placement
-  req.session.privacyCache.set('usingMultiShipping', false);
-
-  // TODO: Exposing a direct route to an Order, without at least encoding the orderID
-  //  is a serious PII violation.  It enables looking up every customers orders, one at a
-  //  time.
-  res.json({
-    error: false,
-    orderID: order.orderNo,
-    orderToken: order.orderToken,
-    continueUrl: URLUtils.url('Order-Confirm').toString()
-  });
-  this.emit('route:Complete', req, res);
-});
+server.prepend('PlaceOrder', server.middleware.https, placeOrder);
 
 module.exports = server.exports();
