+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/checkout_services/__tests__/adyenCheckoutServices.test.js
@@ -1,0 +1,21 @@
+/* eslint-disable global-require */
+let req;
+let res;
+
+beforeEach(() => {
+  jest.clearAllMocks();
+});
+
+describe('isNotAdyen', () => {
+   it('should return false if payment instrument is Adyen', () => {
+       const { isNotAdyen } = require.requireActual('../adyenCheckoutServices');
+       const currentBasket = require('dw/order/BasketMgr').getCurrentBasket();
+        expect(isNotAdyen(currentBasket)).toBeFalsy();
+  });
+    it('should return true if payment instrument is not Adyen', () => {
+        const { isNotAdyen } = require.requireActual('../adyenCheckoutServices');
+        const currentBasket = require('dw/order/BasketMgr').getCurrentBasket();
+        currentBasket.getPaymentInstruments().toArray.mockReturnValue([{paymentMethod: "mockedComponent"}]);
+        expect(isNotAdyen(currentBasket)).toBeTruthy();
+    });
+});
