--- src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/checkout_services/index.js
@@ -1,5 +1,0 @@
-const placeOrder = require('*/cartridge/controllers/middlewares/checkout_services/placeOrder');
-
-module.exports = {
-  placeOrder,
-};
