--- src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/checkout_services/utils/index.js
@@ -1,1 +1,0 @@
-module.exports.createOrder = require('*/cartridge/controllers/middlewares/checkout_services/utils/order');
