--- src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/checkout_services/utils/order.js
@@ -1,135 +1,0 @@
-const Resource = require('dw/web/Resource');
-const URLUtils = require('dw/web/URLUtils');
-const COHelpers = require('*/cartridge/scripts/checkout/checkoutHelpers');
-const hooksHelper = require('*/cartridge/scripts/helpers/hooks');
-const addressHelpers = require('*/cartridge/scripts/helpers/addressHelpers');
-const { fraudDetection } = require('*/cartridge/scripts/hooks/fraudDetection');
-const {
-  hasAdyenPaymentMethod,
-} = require('*/cartridge/controllers/middlewares/checkout_services/helpers/index');
-const handleTransaction = require('*/cartridge/controllers/middlewares/checkout_services/utils/transaction');
-const handlePaymentAuthorization = require('*/cartridge/controllers/middlewares/checkout_services/utils/payment');
-const handleFraudDetection = require('*/cartridge/controllers/middlewares/checkout_services/utils/fraud');
-
-function createOrder(currentBasket, { res, req, next }, emit) {
-  const validateOrder = (order) => {
-    // Creates a new order.
-    if (!order) {
-      res.json({
-        error: true,
-        errorMessage: Resource.msg('error.technical', 'checkout', null),
-      });
-      emit('route:Complete');
-    }
-    return !!order;
-  };
-
-  const handlePlaceOrder = (order, fraudDetectionStatus) => {
-    const placeOrderResult = COHelpers.placeOrder(order, fraudDetectionStatus);
-    if (placeOrderResult.error) {
-      res.json({
-        error: true,
-        errorMessage: Resource.msg('error.technical', 'checkout', null),
-      });
-      emit('route:Complete');
-    }
-
-    return !placeOrderResult.error;
-  };
-
-  const validateOrderAndAuthorize = (order) => {
-    const isValidOrder = validateOrder(order);
-    if (isValidOrder) {
-      const isAuthorized = handlePaymentAuthorization(
-        order,
-        { req, res },
-        emit,
-      );
-      return isAuthorized;
-    }
-    return false;
-  };
-
-  const handleCreateOrder = (order) => {
-    const isAuthorized = validateOrderAndAuthorize(order);
-    if (isAuthorized) {
-      const fraudDetectionStatus = hooksHelper(
-        'app.fraud.detection',
-        'fraudDetection',
-        currentBasket,
-        fraudDetection,
-      );
-      const isSuccessful = handleFraudDetection(
-        fraudDetectionStatus,
-        order,
-        { req, res },
-        emit,
-      );
-
-      // Places the order
-      return isSuccessful && handlePlaceOrder(order, fraudDetectionStatus);
-    }
-    return false;
-  };
-
-  const saveAddresses = ({ currentCustomer }, order) => {
-    if (currentCustomer.addressBook) {
-      const allAddresses = addressHelpers.gatherShippingAddresses(order);
-      allAddresses.forEach((address) => {
-        if (
-          !addressHelpers.checkIfAddressStored(
-            address,
-            currentCustomer.addressBook.addresses,
-          )
-        ) {
-          addressHelpers.saveAddress(
-            address,
-            currentCustomer,
-            addressHelpers.generateAddressName(address),
-          );
-        }
-      });
-    }
-  };
-
-  const isAdyen = hasAdyenPaymentMethod(currentBasket);
-
-  if (!isAdyen) {
-    return next();
-  }
-
-  const isValidTransaction = handleTransaction(
-    currentBasket,
-    { res, req },
-    emit,
-  );
-  if (isValidTransaction) {
-    const order = COHelpers.createOrder(currentBasket);
-
-    saveAddresses(req, order);
-
-    if (order) {
-      // Cache order number in order to be able to restore cart later
-      req.session.privacyCache.set('currentOrderNumber', order.orderNo);
-    }
-
-    const isOrderCreated = handleCreateOrder(order);
-
-    if (isOrderCreated) {
-      COHelpers.sendConfirmationEmail(order, req.locale.id);
-      // Reset usingMultiShip after successful Order placement
-      req.session.privacyCache.set('usingMultiShipping', false);
-
-      res.json({
-        error: false,
-        orderID: order.orderNo,
-        orderToken: order.orderToken,
-        continueUrl: URLUtils.url('Order-Confirm').toString(),
-      });
-      return emit('route:Complete');
-    }
-  }
-  return undefined;
-}
-
-module.exports = createOrder;
